refactor: improve user identification logic in gateway and vault handlers
