+++ internal/vault/handlers/auth.go
