+++ internal/vault/handlers/profile.go
