+++ internal/vault/handlers/search.go
